typing: move DCA types to separate file
